Added jitter tools for Dexter

+++ Dexter/DexterOpcUaTestServer/OpcNodes/BatchParametersNodes.cs
@@ -5,17 +5,17 @@ namespace DexterOpcUaTestServer.OpcNodes
 {
     public class BatchParametersNodes : NodeBase
     {
-        public OpcDataVariableNode<double> FatValue;
-        public OpcDataVariableNode<double> FatTarget;
+        public OpcDataVariableNode<float> FatValue;
+        public OpcDataVariableNode<float> FatTarget;
         public OpcDataVariableNode<string> FatUnit;
-        public OpcDataVariableNode<double> WeightValue;
-        public OpcDataVariableNode<double> WeightTarget;
+        public OpcDataVariableNode<float> WeightValue;
+        public OpcDataVariableNode<float> WeightTarget;
         public OpcDataVariableNode<string> WeightUnit;
-        public OpcDataVariableNode<double> BoneValue;
-        public OpcDataVariableNode<double> BoneTarget;
+        public OpcDataVariableNode<float> BoneValue;
+        public OpcDataVariableNode<float> BoneTarget;
         public OpcDataVariableNode<string> BoneUnit;
-        public OpcDataVariableNode<double> MetalValue;
-        public OpcDataVariableNode<double> MetalTarget;
+        public OpcDataVariableNode<float> MetalValue;
+        public OpcDataVariableNode<float> MetalTarget;
         public OpcDataVariableNode<string> MetalUnit;
         public List<IOpcNode> Nodes => nodes;
         
@@ -44,8 +44,8 @@ namespace DexterOpcUaTestServer.OpcNodes
         {
             var nodes = new List<IOpcNode>();
             var parameterFolder = new OpcFolderNode(parametersFolder, parameterName);
-            var valueNode = CreateOpcUaNode<double>(parameterFolder, $"{parameterName}{NodeSeparator}Result");
-            var targetNode = CreateOpcUaNode<double>(parameterFolder, $"{parameterName}{NodeSeparator}Target");
+            var valueNode = CreateOpcUaNode<float>(parameterFolder, $"{parameterName}{NodeSeparator}Result");
+            var targetNode = CreateOpcUaNode<float>(parameterFolder, $"{parameterName}{NodeSeparator}Target");
             var unitNode = CreateOpcUaNode<string>(parameterFolder, $"{parameterName}{NodeSeparator}Unit");
 
             switch (parameterName)
